Move custom attributable mixin to separate module
